Refactor keyboardSwitch to use notesUtil
