+++ src/components/Key/Key.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import './Key.css';
 
 const Key = ({ note, isSharp, name }) => {
